pyproject.toml
[project]
name = "edgeface-knn"
version = "2.0.0"
description = "Real-time CPU face recognition pipeline using classical ML, refactored from an embedded Raspberry Pi deployment into an installable modular package."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"

authors = [{name = "Saksham Bajaj"}]

keywords = ["face recognition", "computer vision", "opencv", "edge ml", "classical machine learning"]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "opencv-python-headless>=4.5",
    "numpy>=1.21",
    "scikit-learn>=1.0",
    "pyyaml>=5.4",
]

[project.urls]
Homepage = "https://github.com/SakshamBjj/edge-face-recognition-v2"
Source = "https://github.com/SakshamBjj/edge-face-recognition-v2"
Tracker = "https://github.com/SakshamBjj/edge-face-recognition-v2/issues"

[project.scripts]
edge-face = "edge_face.cli:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
